[automated] Merge branch 'master' => 'release/3.0'

+++ src/test/TestUtils/TestArtifact.cs
@@ -1,5 +1,6 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 
 using System;
 using System.Collections.Generic;
@@ -47,7 +48,7 @@ namespace Microsoft.DotNet.CoreSetup.Test
 
         public void Dispose()
         {
-            if (!PreserveTestRuns())
+            if (!PreserveTestRuns() && Directory.Exists(Location))
             {
                 Directory.Delete(Location, true);
             }
